+++ __tests__/lib/mdxish/mdxish.test.ts
@@ -556,7 +556,7 @@ b
 });
 
 describe('html tags rendering', () => {
-  describe('given various attribute formats', () => {
+  describe('various tag attribute formats', () => {
     const expectedAnchor = {
       type: 'element',
       tagName: 'a',
@@ -595,7 +595,7 @@ describe('html tags rendering', () => {
     });
   });
 
-  describe('keeps a non-block level lowercase HTML tags inline', () => {
+  describe('inline-ness vs block-ness of html tags', () => {
     const childTagNames = (parent: Element) =>
       parent.children.filter((c): c is Element => c.type === 'element').map(c => c.tagName);
 
@@ -605,42 +605,44 @@ describe('html tags rendering', () => {
       return paragraphs[0];
     };
 
-    it('keeps a paired icon tag inline with a trailing link', () => {
-      const md =
-        '<i class="fa fa-check fa-lg" style={{ color: "#ff9f3a" }}></i> <a href="https://example.com">3-D Secure</a>';
-      expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
-    });
-
-    it('keeps a self-closing icon tag inline with a trailing link', () => {
-      const md = '<i class="fa fa-check" style={{ color: "red" }}/> <a href="https://example.com">link</a>';
-      expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
-    });
-
-    it('keeps an icon tag inline with trailing plain text', () => {
-      const md = '<i class="fa fa-check" style={{ color: "red" }}></i> done';
-      const paragraph = expectSingleParagraph(mdxish(md));
-      expect(childTagNames(paragraph)).toStrictEqual(['i']);
-      const trailing = paragraph.children.at(-1);
-      expect(trailing).toMatchObject({ type: 'text', value: ' done' });
-    });
-
-    it('keeps the icon inline despite condensed/tab whitespace before the link', () => {
-      const md = '<i style={{ color: "red" }}></i>\t \t<a href="https://example.com">link</a>';
-      expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
-    });
-
-    it('keeps the icon inline inside a list item', () => {
-      const md = '- <i style={{ color: "red" }}></i> <a href="https://example.com">link</a>';
-      const listItem = findElementByTagName(mdxish(md), 'li');
-      expect(listItem).toBeDefined();
-      expect(childTagNames(listItem!)).toStrictEqual(['i', 'a']);
-    });
-
-    it('leaves a standalone icon (no trailing content) as its own block', () => {
-      const md = '<i class="fa fa-check" style={{ color: "red" }}></i>';
-      const paragraphs = mdxish(md).children.filter((c): c is Element => c.type === 'element' && c.tagName === 'p');
-      expect(paragraphs).toHaveLength(0);
-      expect(findElementByTagName(mdxish(md), 'i')).toBeDefined();
+    describe('icon tag', () => {
+      it('keeps a paired icon tag inline with a trailing link', () => {
+        const md =
+          '<i class="fa fa-check fa-lg" style={{ color: "#ff9f3a" }}></i> <a href="https://example.com">3-D Secure</a>';
+        expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
+      });
+  
+      it('keeps a self-closing icon tag inline with a trailing link', () => {
+        const md = '<i class="fa fa-check" style={{ color: "red" }}/> <a href="https://example.com">link</a>';
+        expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
+      });
+  
+      it('keeps an icon tag inline with trailing plain text', () => {
+        const md = '<i class="fa fa-check" style={{ color: "red" }}></i> done';
+        const paragraph = expectSingleParagraph(mdxish(md));
+        expect(childTagNames(paragraph)).toStrictEqual(['i']);
+        const trailing = paragraph.children.at(-1);
+        expect(trailing).toMatchObject({ type: 'text', value: ' done' });
+      });
+  
+      it('keeps the icon inline despite condensed/tab whitespace before the link', () => {
+        const md = '<i style={{ color: "red" }}></i>\t \t<a href="https://example.com">link</a>';
+        expect(childTagNames(expectSingleParagraph(mdxish(md)))).toStrictEqual(['i', 'a']);
+      });
+  
+      it('keeps the icon inline inside a list item', () => {
+        const md = '- <i style={{ color: "red" }}></i> <a href="https://example.com">link</a>';
+        const listItem = findElementByTagName(mdxish(md), 'li');
+        expect(listItem).toBeDefined();
+        expect(childTagNames(listItem!)).toStrictEqual(['i', 'a']);
+      });
+  
+      it('leaves a standalone icon (no trailing content) as its own block', () => {
+        const md = '<i class="fa fa-check" style={{ color: "red" }}></i>';
+        const paragraphs = mdxish(md).children.filter((c): c is Element => c.type === 'element' && c.tagName === 'p');
+        expect(paragraphs).toHaveLength(0);
+        expect(findElementByTagName(mdxish(md), 'i')).toBeDefined();
+      });
     });
 
     // Block-level HTML tags (CommonMark html-block type 6) stay flow even with
@@ -652,57 +654,14 @@ describe('html tags rendering', () => {
       const topLevel = tree.children.filter((c): c is Element => c.type === 'element').map(c => c.tagName);
       expect(topLevel).toStrictEqual(['div', 'p']);
     });
-  });
-  
-  describe('inline components with expression attributes', () => {
-    it('should concatenate expression attributes', () => {
-      const md = "<Anchor label=\"Merchant Sign-Up API Documentation\" target=\"_blank\" href={'https://' + 'www.example.com' + '/reference/sign-up-api'}>Link</Anchor>";
-      const tree = mdxish(md, { newEditorTypes: true });
-      const anchor = findElementByTagName(tree, 'Anchor');
-      expect(anchor).toMatchObject({
-        type: 'element',
-        tagName: 'Anchor',
-        properties: {
-          label: 'Merchant Sign-Up API Documentation',
-          target: '_blank',
-          href: 'https://www.example.com/reference/sign-up-api',
-        },
-        children: [{ type: 'text', value: 'Link' }],
-      });
-    });
 
-    it('should render an Anchor whose href is a concatenation expression', () => {
-      const md =
-        "<Anchor label=\"Docs\" target=\"_blank\" href={'https://' + user.docsUrl + '/x'}>Docs</Anchor>.";
-      const tree = mdxish(md, { newEditorTypes: true });
-      const anchor = findElementByTagName(tree, 'Anchor');
-      expect(anchor).toMatchObject({
-        type: 'element',
-        tagName: 'Anchor',
-        properties: {
-          label: 'Docs',
-          target: '_blank',
-          href: "'https://' + user.docsUrl + '/x'",
-        },
-        children: [{ type: 'text', value: 'Docs' }],
-      });
-    });
-
-    it('should keep trailing text after the closing tag as a sibling', () => {
-      const md = "Start <Anchor href={'a' + 'b'}>Link</Anchor> done.";
-      const tree = mdxish(md, { newEditorTypes: true });
-      const paragraph = tree.children[0] as Element;
-      const [start, anchor, trailing] = paragraph.children;
-      expect(start).toMatchObject({ type: 'text', value: 'Start ' });
-      expect(anchor).toMatchObject({
-        type: 'element',
-        tagName: 'Anchor',
-        properties: {
-          href: 'ab',
-        },
-        children: [{ type: 'text', value: 'Link' }],
-      });
-      expect(trailing).toMatchObject({ type: 'text', value: ' done.' });
+    // Raw-content tags (e.g. <pre>) also stay flow
+    it('keeps a raw-content tag as its own block with trailing content as a sibling', () => {
+      const md = '<pre data={{ a: 1 }}>code</pre> trailing';
+      const tree = mdxish(md);
+      expect(findElementByTagName(tree, 'pre')).toBeDefined();
+      const topLevel = tree.children.filter((c): c is Element => c.type === 'element').map(c => c.tagName);
+      expect(topLevel).toStrictEqual(['pre', 'p']);
     });
   });
 });
